backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
PyGithub==2.1.1
GitPython==3.1.40
faiss-cpu==1.11.0
rank-bm25==0.2.2
sentence-transformers==2.2.2
llama-cpp-python==0.2.20
langchain-text-splitters==0.0.1
pydantic==2.9.0
python-dotenv==1.0.0
aiofiles==23.2.1
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
websockets==12.0